requirements.txt
# Core Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17

# Chemistry & Molecular Processing
rdkit>=2023.3.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
optuna>=3.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Configuration & Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Experiment Tracking
mlflow>=2.4.0
wandb>=0.15.0
tensorboard>=2.13.0

# API & Deployment
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Model Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Chemistry Visualization
py3Dmol>=2.0.0
